+++ google-beta/data_source_dns_record_set_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 	"github.com/hashicorp/terraform-provider-google-beta/google-beta/acctest"
+	"github.com/hashicorp/terraform-provider-google-beta/google-beta/fwtransport"
 )
 
 func TestAccDataSourceDnsRecordSet_basic(t *testing.T) {
@@ -35,7 +36,7 @@ func TestAccDataSourceDnsRecordSet_basic(t *testing.T) {
 				Config: testAccDataSourceDnsRecordSet_basic(managedZoneName, RandString(t, 10)),
 				Check: resource.ComposeTestCheckFunc(
 					acctest.CheckDataSourceStateMatchesResourceState("data.google_dns_record_set.rs", "google_dns_record_set.rs"),
-					testExtractResourceAttr("data.google_dns_record_set.rs", "ttl", &ttl1),
+					acctest.TestExtractResourceAttr("data.google_dns_record_set.rs", "ttl", &ttl1),
 				),
 			},
 			{
@@ -43,8 +44,8 @@ func TestAccDataSourceDnsRecordSet_basic(t *testing.T) {
 				Config:                   testAccDataSourceDnsRecordSet_basic(managedZoneName, RandString(t, 10)),
 				Check: resource.ComposeTestCheckFunc(
 					acctest.CheckDataSourceStateMatchesResourceState("data.google_dns_record_set.rs", "google_dns_record_set.rs"),
-					testExtractResourceAttr("data.google_dns_record_set.rs", "ttl", &ttl2),
-					testCheckAttributeValuesEqual(&ttl1, &ttl2),
+					acctest.TestExtractResourceAttr("data.google_dns_record_set.rs", "ttl", &ttl2),
+					acctest.TestCheckAttributeValuesEqual(&ttl1, &ttl2),
 				),
 			},
 		},
@@ -90,18 +91,18 @@ func testAccCheckDnsRecordSetDestroyProducerFramework(t *testing.T) func(s *terr
 
 			p := GetFwTestProvider(t)
 
-			url, err := replaceVarsForFrameworkTest(&p.frameworkProvider, rs, "{{DNSBasePath}}projects/{{project}}/managedZones/{{managed_zone}}/rrsets/{{name}}/{{type}}")
+			url, err := acctest.ReplaceVarsForFrameworkTest(&p.FrameworkProvider.FrameworkProviderConfig, rs, "{{DNSBasePath}}projects/{{project}}/managedZones/{{managed_zone}}/rrsets/{{name}}/{{type}}")
 			if err != nil {
 				return err
 			}
 
 			billingProject := ""
 
-			if !p.billingProject.IsNull() && p.billingProject.String() != "" {
-				billingProject = p.billingProject.String()
+			if !p.BillingProject.IsNull() && p.BillingProject.String() != "" {
+				billingProject = p.BillingProject.String()
 			}
 
-			_, diags := sendFrameworkRequest(&p.frameworkProvider, "GET", billingProject, url, p.userAgent, nil)
+			_, diags := fwtransport.SendFrameworkRequest(&p.FrameworkProvider.FrameworkProviderConfig, "GET", billingProject, url, p.UserAgent, nil)
 			if !diags.HasError() {
 				return fmt.Errorf("DNSResourceDnsRecordSet still exists at %s", url)
 			}
