+++ google-beta/framework_test_utils.go
@@ -3,15 +3,11 @@
 package google
 
 import (
-	"fmt"
-	"reflect"
-	"regexp"
-	"strings"
-
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 
-	"github.com/hashicorp/terraform-provider-google-beta/google-beta/tpgresource"
+	"github.com/hashicorp/terraform-provider-google-beta/google-beta/acctest"
+	"github.com/hashicorp/terraform-provider-google-beta/google-beta/fwtransport"
 )
 
 // General test utils
@@ -19,86 +15,16 @@ import (
 // testExtractResourceAttr navigates a test's state to find the specified resource (or data source) attribute and makes the value
 // accessible via the attributeValue string pointer.
 func testExtractResourceAttr(resourceName string, attributeName string, attributeValue *string) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		rs, ok := s.RootModule().Resources[resourceName] // To find a datasource, include `data.` at the start of the resourceName value
-
-		if !ok {
-			return fmt.Errorf("resource name %s not found in state", resourceName)
-		}
-
-		attrValue, ok := rs.Primary.Attributes[attributeName]
-
-		if !ok {
-			return fmt.Errorf("attribute %s not found in resource %s state", attributeName, resourceName)
-		}
-
-		*attributeValue = attrValue
-
-		return nil
-	}
+	return acctest.TestExtractResourceAttr(resourceName, attributeName, attributeValue)
 }
 
 // testCheckAttributeValuesEqual compares two string pointers, which have been used to retrieve attribute values from the test's state.
 func testCheckAttributeValuesEqual(i *string, j *string) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		if testStringValue(i) != testStringValue(j) {
-			return fmt.Errorf("attribute values are different, got %s and %s", testStringValue(i), testStringValue(j))
-		}
-
-		return nil
-	}
-}
-
-// testStringValue returns string values from string pointers, handling nil pointers.
-func testStringValue(sPtr *string) string {
-	if sPtr == nil {
-		return ""
-	}
-
-	return *sPtr
+	return acctest.TestCheckAttributeValuesEqual(i, j)
 }
 
 // This function isn't a test of transport.go; instead, it is used as an alternative
 // to ReplaceVars inside tests.
-func replaceVarsForFrameworkTest(prov *frameworkProvider, rs *terraform.ResourceState, linkTmpl string) (string, error) {
-	re := regexp.MustCompile("{{([[:word:]]+)}}")
-	var project, region, zone string
-
-	if strings.Contains(linkTmpl, "{{project}}") {
-		project = rs.Primary.Attributes["project"]
-	}
-
-	if strings.Contains(linkTmpl, "{{region}}") {
-		region = tpgresource.GetResourceNameFromSelfLink(rs.Primary.Attributes["region"])
-	}
-
-	if strings.Contains(linkTmpl, "{{zone}}") {
-		zone = tpgresource.GetResourceNameFromSelfLink(rs.Primary.Attributes["zone"])
-	}
-
-	replaceFunc := func(s string) string {
-		m := re.FindStringSubmatch(s)[1]
-		if m == "project" {
-			return project
-		}
-		if m == "region" {
-			return region
-		}
-		if m == "zone" {
-			return zone
-		}
-
-		if v, ok := rs.Primary.Attributes[m]; ok {
-			return v
-		}
-
-		// Attempt to draw values from the provider
-		if f := reflect.Indirect(reflect.ValueOf(prov)).FieldByName(m); f.IsValid() {
-			return f.String()
-		}
-
-		return ""
-	}
-
-	return re.ReplaceAllStringFunc(linkTmpl, replaceFunc), nil
+func replaceVarsForFrameworkTest(prov *fwtransport.FrameworkProviderConfig, rs *terraform.ResourceState, linkTmpl string) (string, error) {
+	return acctest.ReplaceVarsForFrameworkTest(prov, rs, linkTmpl)
 }
