+++ google-beta/framework_utils.go
@@ -4,93 +4,38 @@ package google
 
 import (
 	"context"
-	"fmt"
-	"os"
-	"regexp"
-	"strings"
 
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
 	"github.com/hashicorp/terraform-plugin-framework/types"
-	"github.com/hashicorp/terraform-plugin-log/tflog"
+
+	"github.com/hashicorp/terraform-provider-google-beta/google-beta/fwmodels"
+	"github.com/hashicorp/terraform-provider-google-beta/google-beta/fwresource"
+	"github.com/hashicorp/terraform-provider-google-beta/google-beta/fwtransport"
 	"github.com/hashicorp/terraform-provider-google-beta/google-beta/tpgresource"
-	transport_tpg "github.com/hashicorp/terraform-provider-google-beta/google-beta/transport"
 )
 
-const uaEnvVar = "TF_APPEND_USER_AGENT"
-
 func CompileUserAgentString(ctx context.Context, name, tfVersion, provVersion string) string {
-	ua := fmt.Sprintf("Terraform/%s (+https://www.terraform.io) Terraform-Plugin-SDK/%s %s/%s", tfVersion, "terraform-plugin-framework", name, provVersion)
-
-	if add := os.Getenv(uaEnvVar); add != "" {
-		add = strings.TrimSpace(add)
-		if len(add) > 0 {
-			ua += " " + add
-			tflog.Debug(ctx, fmt.Sprintf("Using modified User-Agent: %s", ua))
-		}
-	}
-
-	return ua
+	return fwtransport.CompileUserAgentString(ctx, name, tfVersion, provVersion)
 }
 
-func GetCurrentUserEmailFramework(p *frameworkProvider, userAgent string, diags *diag.Diagnostics) string {
-	// When environment variables UserProjectOverride and BillingProject are set for the provider,
-	// the header X-Goog-User-Project is set for the API requests.
-	// But it causes an error when calling GetCurrUserEmail. Set the project to be "NO_BILLING_PROJECT_OVERRIDE".
-	// And then it triggers the header X-Goog-User-Project to be set to empty string.
-
-	// See https://github.com/golang/oauth2/issues/306 for a recommendation to do this from a Go maintainer
-	// URL retrieved from https://accounts.google.com/.well-known/openid-configuration
-	res, d := sendFrameworkRequest(p, "GET", "NO_BILLING_PROJECT_OVERRIDE", "https://openidconnect.googleapis.com/v1/userinfo", userAgent, nil)
-	diags.Append(d...)
-
-	if diags.HasError() {
-		tflog.Info(p.context, "error retrieving userinfo for your provider credentials. have you enabled the 'https://www.googleapis.com/auth/userinfo.email' scope?")
-		return ""
-	}
-	if res["email"] == nil {
-		diags.AddError("error retrieving email from userinfo.", "email was nil in the response.")
-		return ""
-	}
-	return res["email"].(string)
+func GetCurrentUserEmailFramework(p *fwtransport.FrameworkProviderConfig, userAgent string, diags *diag.Diagnostics) string {
+	return fwtransport.GetCurrentUserEmailFramework(p, userAgent, diags)
 }
 
-func generateFrameworkUserAgentString(metaData *ProviderMetaModel, currUserAgent string) string {
-	if metaData != nil && !metaData.ModuleName.IsNull() && metaData.ModuleName.ValueString() != "" {
-		return strings.Join([]string{currUserAgent, metaData.ModuleName.ValueString()}, " ")
-	}
-
-	return currUserAgent
+func generateFrameworkUserAgentString(metaData *fwmodels.ProviderMetaModel, currUserAgent string) string {
+	return fwtransport.GenerateFrameworkUserAgentString(metaData, currUserAgent)
 }
 
 // GetProject reads the "project" field from the given resource and falls
 // back to the provider's value if not given. If the provider's value is not
 // given, an error is returned.
 func getProjectFramework(rVal, pVal types.String, diags *diag.Diagnostics) types.String {
-	return getProjectFromFrameworkSchema("project", rVal, pVal, diags)
-}
-
-func getProjectFromFrameworkSchema(projectSchemaField string, rVal, pVal types.String, diags *diag.Diagnostics) types.String {
-	if !rVal.IsNull() && rVal.ValueString() != "" {
-		return rVal
-	}
-
-	if !pVal.IsNull() && pVal.ValueString() != "" {
-		return pVal
-	}
-
-	diags.AddError("required field is not set", fmt.Sprintf("%s is not set", projectSchemaField))
-	return types.String{}
+	return fwresource.GetProjectFramework(rVal, pVal, diags)
 }
 
 func handleDatasourceNotFoundError(ctx context.Context, err error, state *tfsdk.State, resource string, diags *diag.Diagnostics) {
-	if transport_tpg.IsGoogleApiErrorWithCode(err, 404) {
-		tflog.Warn(ctx, fmt.Sprintf("Removing %s because it's gone", resource))
-		// The resource doesn't exist anymore
-		state.RemoveResource(ctx)
-	}
-
-	diags.AddError(fmt.Sprintf("Error when reading or editing %s", resource), err.Error())
+	fwtransport.HandleDatasourceNotFoundError(ctx, err, state, resource, diags)
 }
 
 // field helpers
@@ -98,33 +43,5 @@ func handleDatasourceNotFoundError(ctx context.Context, err error, state *tfsdk.
 // Parses a project field with the following formats:
 // - projects/{my_projects}/{resource_type}/{resource_name}
 func parseProjectFieldValueFramework(resourceType, fieldValue, projectSchemaField string, rVal, pVal types.String, isEmptyValid bool, diags *diag.Diagnostics) *tpgresource.ProjectFieldValue {
-	if len(fieldValue) == 0 {
-		if isEmptyValid {
-			return &tpgresource.ProjectFieldValue{ResourceType: resourceType}
-		}
-		diags.AddError("field can not be empty", fmt.Sprintf("The project field for resource %s cannot be empty", resourceType))
-		return nil
-	}
-
-	r := regexp.MustCompile(fmt.Sprintf(tpgresource.ProjectBasePattern, resourceType))
-	if parts := r.FindStringSubmatch(fieldValue); parts != nil {
-		return &tpgresource.ProjectFieldValue{
-			Project: parts[1],
-			Name:    parts[2],
-
-			ResourceType: resourceType,
-		}
-	}
-
-	project := getProjectFromFrameworkSchema(projectSchemaField, rVal, pVal, diags)
-	if diags.HasError() {
-		return nil
-	}
-
-	return &tpgresource.ProjectFieldValue{
-		Project: project.ValueString(),
-		Name:    tpgresource.GetResourceNameFromSelfLink(fieldValue),
-
-		ResourceType: resourceType,
-	}
+	return fwresource.ParseProjectFieldValueFramework(resourceType, fieldValue, projectSchemaField, rVal, pVal, isEmptyValid, diags)
 }
